feat(docs): single agent-led landing page
