+++ docs/app/pages/index.vue
@@ -6,10 +6,17 @@ useHead({ titleTemplate: () => "ViteHub" });
 useSeoMeta({
   title: "ViteHub — Agents for any host",
   ogTitle: "ViteHub — Agents for any host · ViteHub",
-  description: "Compose agents with server primitives for databases, storage, workflows, schedules, and workspaces through one Vite plugin.",
+  description: "Define any agent, compose it from capabilities, and ship it to any host — built on server primitives, wired with one Vite plugin.",
 });
 </script>
 
 <template>
-  <LandingHero />
+  <UMain class="bg-default text-default">
+    <LandingHero />
+    <LandingAnyHost />
+    <LandingAnatomy />
+    <LandingPrimitives />
+    <LandingShowcase />
+    <LandingCta />
+  </UMain>
 </template>
